refactor:서버 컴포넌트의 불필요한 API 호출 제거 및 전시회 로직 분리
