+++ src/app/api/exhibitions/route.ts
@@ -4,162 +4,27 @@ import {
   parseFormDataToObj,
   validateExhibition,
 } from '@/components/galleryExhibition/threejs/test/util/util';
-import { ExhibitionListItem, ExhibitionRow } from '@/types/exhibitionList';
-import { EXHIBITIONS_PER_PAGE } from '@/lib/exhibition/constants';
-
-// 한국 시간 기준 오늘 날짜 (YYYY-MM-DD)
-const todayKST = (): string => {
-  const kst = new Date(Date.now() + 9 * 60 * 60 * 1000);
-  return kst.toISOString().slice(0, 10);
-};
+import {
+  ExhibitionsAuthRequiredError,
+  fetchExhibitions,
+} from '@/lib/exhibition/queries';
 
 export async function GET(req: NextRequest) {
   try {
-    const supabase = await createClient();
-    const today = todayKST();
-    const sort = req.nextUrl.searchParams.get('sort');
+    const search = req.nextUrl.searchParams.get('search') || '';
+    const sort = req.nextUrl.searchParams.get('sort') || 'latest';
+    const rawPage = parseInt(req.nextUrl.searchParams.get('page') || '1', 10);
+    const page = Number.isFinite(rawPage) && rawPage > 0 ? rawPage : 1;
 
-    // 검색 파라미터
-    const search = req.nextUrl.searchParams.get('search');
+    const { data, pagination } = await fetchExhibitions({ page, sort, search });
 
-    // pagination
-    const page = parseInt(req.nextUrl.searchParams.get('page') || '1', 10);
-    const limit = parseInt(
-      req.nextUrl.searchParams.get('limit') || String(EXHIBITIONS_PER_PAGE),
-      10
-    );
-    const from = (page - 1) * limit;
-    const to = from + limit - 1;
-
-    let query = supabase.from('exhibitions').select(
-      `
-      id,
-      title,
-      thumbnail_url,
-      start_date,
-      end_date,
-      teacher_id,
-      created_at,
-      profile:profiles!teacher_id ( institution ),
-      likes:exhibition_likes ( count )
-    `,
-      { count: 'exact' }
-    );
-
-    // 검색
-    if (search) {
-      const { data: matchedProfiles } = await supabase
-        .from('profiles')
-        .select('id')
-        .ilike('institution', `%${search}%`);
-
-      const matchedTeacherIds =
-        matchedProfiles?.map((profile) => profile.id) || [];
-
-      if (matchedTeacherIds.length > 0) {
-        const idsString = `(${matchedTeacherIds.join(',')})`;
-        query = query.or(`title.ilike.%${search}%,teacher_id.in.${idsString}`);
-      } else {
-        query = query.ilike('title', `%${search}%`);
-      }
-    }
-
-    // 필터링
-    switch (sort) {
-      case 'oldest':
-        query = query
-          .lte('start_date', today)
-          .or(`end_date.gte.${today},end_date.is.null`)
-          .order('start_date', { ascending: true });
-        break;
-      case 'popular':
-        query = query
-          .lte('start_date', today)
-          .or(`end_date.gte.${today},end_date.is.null`)
-          .order('start_date', { ascending: false });
-        break;
-      case 'upcoming':
-        query = query
-          .gt('start_date', today)
-          .order('start_date', { ascending: true });
-        break;
-      case 'ended':
-        query = query
-          .lt('end_date', today)
-          .order('end_date', { ascending: false });
-        break;
-      case 'mine': {
-        const {
-          data: { user },
-        } = await supabase.auth.getUser();
-        if (!user) {
-          return NextResponse.json({ message: 'no session' }, { status: 401 });
-        }
-        query = query
-          .eq('teacher_id', user.id)
-          .order('created_at', { ascending: false });
-        break;
-      }
-
-      case 'latest':
-      default:
-        query = query
-          .lte('start_date', today)
-          .or(`end_date.gte.${today},end_date.is.null`)
-          .order('start_date', { ascending: false });
-    }
-
-    // TODO: popular 정렬은 추후 새로운 이슈에서 좋아요 작업할때 수정예정
-    // (현재는 range를 건너뛰고 메모리 정렬 → 1000건 상한 및 페이지네이션 미동작)
-    if (sort !== 'popular') {
-      query = query.range(from, to);
-    }
-
-    const { data, error, count } = await query.returns<ExhibitionRow[]>();
-
-    if (error) {
-      if (error.code === 'PGRST103') {
-        return NextResponse.json(
-          {
-            data: [],
-            pagination: { page, limit, totalCount: 0, hasNextPage: false },
-          },
-          { status: 200 }
-        );
-      }
-
-      console.log(error);
-      return NextResponse.json({ message: 'database error' }, { status: 500 });
-    }
-
-    let result: ExhibitionListItem[] = (data ?? []).map((row) => {
-      const profile = Array.isArray(row.profile) ? row.profile[0] : row.profile;
-
-      return {
-        id: row.id,
-        title: row.title,
-        host: profile?.institution ?? '',
-        image: row.thumbnail_url,
-        startDate: row.start_date,
-        endDate: row.end_date,
-        likes: row.likes?.[0]?.count ?? 0,
-      };
-    });
-
-    // 인기순 재정렬
-    if (sort === 'popular') {
-      result = result.sort((a, b) => b.likes - a.likes);
-    }
-
-    const totalCount = count || 0;
-    const hasNextPage = from + limit < totalCount;
-
-    return NextResponse.json(
-      { data: result, pagination: { page, limit, totalCount, hasNextPage } },
-      { status: 200 }
-    );
+    return NextResponse.json({ data, pagination }, { status: 200 });
   } catch (err) {
     console.error(err);
+    // sort = mine 미로그인시 에러
+    if (err instanceof ExhibitionsAuthRequiredError) {
+      return NextResponse.json({ message: 'no session' }, { status: 401 });
+    }
     return NextResponse.json({ message: 'unknown error' }, { status: 500 });
   }
 }
